big update

+++ server/Domain/Entities/ProductAttribute.cs
@@ -9,8 +9,8 @@ public class ProductAttribute
     public string Value { get; set; } = string.Empty;
     public string ProductId { get; set; } = string.Empty;
     public Product? Product { get; set; }
-    public int DisplayOrder { get; set; } 
-    public string AttributeType { get; set; } = string.Empty; // "text", "number", "color", "size", etc.
+    public long DisplayOrder { get; set; } 
+    public string AttributeType { get; set; } = string.Empty; 
 }
 
 public enum AttributeType
@@ -18,6 +18,5 @@ public enum AttributeType
     ProcessorAndGraphics,
     MemoryAndStorage,
     Dispay,
-    Audio,
-    
+    Audio,    
 }
